Modified API project structure
